requirement.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-community
langchain-google-genai
langchain-qdrant
qdrant-client
sentence-transformers
langchain-classic
regex
